Change default status for --prop:/sim/mooring to false
Now, when launching, it will come for hard-surface runway by default.
The advantage is that now, you can set any lat-lon-heading location,
and set it to moor as expected by adding the property to true

--prop:/sim/mooring=true

Doing this, will also execute the code to moor appropriately in airports with
indicated mooring locations.